Rename PostText to PieceOfText
Since it's used both for posts and comments
