+++ src/components/piece-of-text.jsx
